# PaymentRecoveryAPI/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Razopay Agent API",
    "description": "Development Environment.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/onboard-business/": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:Initiate",
        "description": "Step 1 of onboarding.\n\nRegisters a `PENDING` business row and returns the Razorpay `authorization_url`.\nRedirect the business owner there; Razorpay calls back to `/onboard-business/callback`.",
        "operationId": "onboarding_initiate_api_onboard_business__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessOnboardRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessOnboardInitResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:List",
        "operationId": "onboarding_list_api_onboard_business__get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BusinessResponse"
                  },
                  "title": "Response Onboarding List Api Onboard Business  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboard-business/callback": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:Callback",
        "description": "Step 2 of onboarding (Razorpay OAuth redirect target).\n\nExchanges `code` for tokens, stores them encrypted, then creates the\nsub-merchant webhook for the configured event triggers.",
        "operationId": "onboarding_callback_api_onboard_business_callback_get",
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "State"
            }
          },
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "error",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboard-business/{business_id}": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:Get-By-Id",
        "operationId": "onboarding_get_by_id_api_onboard_business__business_id__get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboard-business/{business_id}/recovery-cases": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:List-Recovery-Cases",
        "description": "Every merged recovery case for this business, most recently active first.",
        "operationId": "onboarding_list_recovery_cases_api_onboard_business__business_id__recovery_cases_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RecoveryCaseResponse"
                  },
                  "title": "Response Onboarding List Recovery Cases Api Onboard Business  Business Id  Recovery Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboard-business/{business_id}/webhooks": {
      "get": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:Get-Webhook-Config",
        "description": "Live config (url, active, events) of the webhook Razorpay has registered for\nthis business - fetched from Razorpay on every call, not a local cache.",
        "operationId": "onboarding_get_webhook_config_api_onboard_business__business_id__webhooks_get",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/onboard-business/{business_id}/refresh-token": {
      "post": {
        "tags": [
          "onboarding"
        ],
        "summary": "Onboarding:Refresh-Token",
        "operationId": "onboarding_refresh_token_api_onboard_business__business_id__refresh_token_post",
        "parameters": [
          {
            "name": "business_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Business Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BusinessOnboardInitResponse": {
        "properties": {
          "businessId": {
            "type": "integer",
            "title": "Businessid"
          },
          "referenceId": {
            "type": "string",
            "title": "Referenceid"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "authorizationUrl": {
            "type": "string",
            "title": "Authorizationurl"
          },
          "state": {
            "type": "string",
            "title": "State"
          }
        },
        "type": "object",
        "required": [
          "businessId",
          "referenceId",
          "status",
          "authorizationUrl",
          "state"
        ],
        "title": "BusinessOnboardInitResponse",
        "description": "Returned by `POST /onboard-business/` - redirect the business owner here to grant access."
      },
      "BusinessOnboardRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "referenceId": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Referenceid"
          },
          "contactEmail": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contactemail"
          },
          "scope": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Scope"
          }
        },
        "type": "object",
        "required": [
          "name",
          "referenceId"
        ],
        "title": "BusinessOnboardRequest",
        "description": "Payload to kick off onboarding of a new sub-merchant business."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "WebhookConfigResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "active": {
            "type": "boolean",
            "title": "Active"
          },
          "events": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Events"
          },
          "alert_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alert Email"
          },
          "secret_exists": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Exists"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "active",
          "events"
        ],
        "title": "WebhookConfigResponse",
        "description": "The live config of a business' registered Razorpay webhook (not cached)."
      },
...
    }
  }
}